linting vetting formatting idiomating etc

+++ project.go
@@ -9,7 +9,12 @@ import (
 func Ortho(left, right, bottom, top, near, far float32) Mat4 {
 	rml, tmb, fmn := (right - left), (top - bottom), (far - near)
 
-	return Mat4{float32(2. / rml), 0, 0, 0, 0, float32(2. / tmb), 0, 0, 0, 0, float32(-2. / fmn), 0, float32(-(right + left) / rml), float32(-(top + bottom) / tmb), float32(-(far + near) / fmn), 1}
+	return Mat4{
+		2 / rml, 0, 0, 0,
+		0, 2 / tmb, 0, 0,
+		0, 0, -2 / fmn, 0,
+		-(right + left) / rml, -(top + bottom) / tmb, -(far + near) / fmn, 1,
+	}
 }
 
 // Ortho2D is equivalent to Ortho with the near and far planes being -1 and 1, respectively
@@ -19,10 +24,14 @@ func Ortho2D(left, right, bottom, top float32) Mat4 {
 
 // Perspective returns a Mat4 representing a perspective projection of the given arguments.
 func Perspective(fovy, aspect, near, far float32) Mat4 {
-	// fovy = (fovy * math.Pi) / 180.0 // convert from degrees to radians
 	nmf, f := near-far, 1./math.Tan(fovy/2.0)
 
-	return Mat4{float32(f / aspect), 0, 0, 0, 0, float32(f), 0, 0, 0, 0, float32((near + far) / nmf), -1, 0, 0, float32((2. * far * near) / nmf), 0}
+	return Mat4{
+		f / aspect, 0, 0, 0,
+		0, f, 0, 0, 0,
+		0, (near + far) / nmf, -1, 0,
+		0, (2. * far * near) / nmf, 0,
+	}
 }
 
 // Frustum returns a Mat4 representing a frustrum transform (squared pyramid with the top cut off)
@@ -30,7 +39,12 @@ func Frustum(left, right, bottom, top, near, far float32) Mat4 {
 	rml, tmb, fmn := (right - left), (top - bottom), (far - near)
 	A, B, C, D := (right+left)/rml, (top+bottom)/tmb, -(far+near)/fmn, -(2*far*near)/fmn
 
-	return Mat4{float32((2. * near) / rml), 0, 0, 0, 0, float32((2. * near) / tmb), 0, 0, float32(A), float32(B), float32(C), -1, 0, 0, float32(D), 0}
+	return Mat4{
+		(2. * near) / rml, 0, 0, 0,
+		0, (2. * near) / tmb, 0, 0,
+		A, B, C, -1,
+		0, 0, D, 0,
+	}
 }
 
 // LookAt returns a Mat4 that represents a camera transform from the given arguments.
